project_binance_arbitrage_tracker/requirements.txt
Flask==0.12.2
pandas==0.22.0
python-binance==0.6.3
XlsxWriter==1.0.2
simplejson==3.13.2